Feat : Edit Archive and Category Entity, Change Id value type int to Long

+++ src/main/java/com/weave/weaveserver/dto/ArchiveResponse.java
@@ -10,7 +10,7 @@ public class ArchiveResponse {
     @AllArgsConstructor
     @Data
     public static class listResponse{
-        private int categoryIdx; //임시
+        private Long categoryIdx; //임시
 
     }
 
@@ -18,7 +18,7 @@ public class ArchiveResponse {
     @AllArgsConstructor
     @Data
     public static class detailResponse{
-        private int archiveIdx; //임시
+        private Long archiveIdx; //임시
 
     }
