+++ src/main/java/com/weave/weaveserver/repository/ArchiveRepository.java
@@ -4,7 +4,7 @@ import com.weave.weaveserver.domain.Archive;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ArchiveRepository extends JpaRepository<Archive, Integer> {
+public interface ArchiveRepository extends JpaRepository<Archive, Long> {
 //    @Query(value = "SELECT ")
     //TODO : 작성
